% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_bootstrap_datasets.R
\name{boot_f}
\alias{boot_f}
\title{Generating bootstrap sample datasets}
\usage{
boot_f(df = NA, nboot = 100, boot_ft = 5, seed = 20231106)
}
\arguments{
\item{df}{working dataset,simulated or real data collected}

\item{nboot}{number of bootstrap datasets}

\item{boot_ft}{number of features in each bootstrap dataset}

\item{seed}{random seed}
}
\value{
a list, size = nboot, each element is a dataframe
}
\description{
Generating bootstrap sample datasets
}
\examples{
library(survival)
dat <- sim_survdat_f(nsample =1000, varnum = 25,dist='g',lambda=0.01, rho=1, beta=c(2.3,0.3,0.4), crate=0.001,cor=TRUE,seed=20231106)
fit <- survival::coxph(Surv(time, status) ~ X1+X3+X1*X3, data=dat)
boot_lst <- boot_f(df = ,nboot = 100,boot_ft = 5,seed = 20231106)
}
