% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_simulate_survival_data.R
\name{sim_survdat_f}
\alias{sim_survdat_f}
\title{Simulate a survival data set}
\usage{
sim_survdat_f(
  nsample = 100,
  varnum = 25,
  dist = "w",
  lambda = 0.01,
  rho = 1,
  beta = NA,
  crate = 0.001,
  cor = TRUE,
  seed = 20231106,
  ...
)
}
\arguments{
\item{nsample}{total number of observations}

\item{varnum}{total number of covariates}

\item{dist}{distribution of baseline hazard function}

\item{lambda}{scale parameter of the distribution}

\item{rho}{shape parameter of the distribution}

\item{beta}{pre-specified effect size of covariates}

\item{crate}{rate parameter of the exponential distribution}

\item{cor}{logical; whether consider the correlation between covariates}

\item{seed}{}
}
\value{
a dataframe; time, status, covariates
}
\description{
Simulate a survival data set
}
\examples{
library(survival)
dat <- sim_survdat_f(nsample =1000, varnum = 25,dist='g',lambda=0.01, rho=1, beta=c(2.3,0.3,0.4), crate=0.001,cor=TRUE,seed=20231106)
fit <- survival::coxph(Surv(time, status) ~ X1+X3+X1*X3, data=dat)
}
